Rework dataset model and file dragging. Use single model for dragging, uploading, and selecting images.
